Eliminate use of manual memory management
This code contained a bug, mixing new[] with delete instead of delete[].
This could potentially cause memory corruption under some C++ runtimes.

Using `std::string` is simpler. Buffer space can be allocated with
`reserve`, though this does not update the size of the string. Buffer
space can also be allocated with `resize`, which will update the size of
the string, though also at the cost of initializing the buffer with
nulls. This is largely wasted initialization time since the buffer is
then immediately filled with data from the file. There does not appear
to be a standard library method to update the size of a `std::string`
without also initializing that length of the buffer.

Read of file data should use a fixed sized buffer for the length.

Use of &s[0] makes use of C++11 standard which states string buffer is
stored contiguously.