Convert MapReader/Writer stream pointers to ref
This allows the . syntax as opposed to the -> syntax.

The class structure is odd though, in that there is effectively a single
public member function, which can only be called once per object. This
would make more sense as a free function in a namespace.

Current code has a warning due to no constructor, which leave the
internal reference unitialized after object construction. It is
initialized at the start of the public method.