Reorganize StreamReader/StreamWriter interfaces
This places helper methods at the bottom, which depend only on the core
interface above.

Side note: As a micro optimization, putting the most common virtual
method first (ReadImplementation/WriteImplementation) means the
method will appear first in the virtual function table. This means the
method is access with an offset of 0, and instructions with no offset
have a smaller encoding. Hence, calls to this method will have a
slightly more compact instruction encoding.

Organizational note: This also places the core method for this interface
first, giving it the prominence it deserves.