Set CMake standard required setting
Without the standard required setting, CMake can try to fall back to previous versions of C++. We explicitly use C++14 features though, so the code would fail to compile under earlier standards.

The reason for this setting seems to be the default standard on Linux is an older one, with new features and updates needing to be explicitly enabled. Hence code does not compile using the latest updates unless it is explicitly requested. This does not always mean code needs the latest features to compile.

Also modified the spacing on `set` to match other instances.