Implement length prefix vector read
Curiously, the implementation is exactly the same as the templated
length prefixed string read. Perhaps there is a way to make the template
parameter more generic and provide both with one source method.

One complication with combining length prefixed vector and string
methods is they both contain an implicit assumption of contiguous
memory. This assumption comes from the `Read` on the last line of both
methods. Although this assumption is guaranteed by the standard for both
`vector` (since C++03), and `string` (since C++11), it is not true for
other more generic STL containers. Hence combining code to one method
would need to restrict the method to those two data types to prevent bad
code generation for other STL containers.

For now though, this works, and closes #133.