Add templated string write method
Allows writing strings without writing a size prefix.

This could be used for binary writes where the string length is
determined by some external means. Perhaps a length prefix is given
earlier in the file as part of a header, is somehow calculated, or is
known to be of a fixed constant length. This method can also be used
with text files.