Add templated string read method
Allows reading strings without a size prefix.

The size of the string is determined by some external means, and must be
manually sized before the read operation. Perhaps a length is given as
part of a header field located elsewhere in the file, is somehow
calculated, or known to be of a fixed constant length. This could be
used to read from text files, assuming an appropriate read size can be
predetermined.