Rename method for clarity
We are getting an index into the `tiles` vector, with `TileData`
entries. A "Cell" is a field within a `TileData` struct, so the previous
name was a bit confusing.

Granted, there is still the distinction between TileData, and the index
of the actual tile graphic within all loaded tilesets, and the index of
the tile graphic within a specific tileset. Perhaps some additional
renaming of those fields is needed for improved clarity.