Use `std::size_t` for memory index
I was thinking of leaving this, since it might make updates easier to
keep the same data type, but then I realized we need to bounds check
writes anyway, and so a different type will remind us of that.

Additionally, using `std::size_t` means the interface here is not
limited to the same restrictions as the OP2 map file format. If we
wanted to support larger maps in an alternate format someday, we could
do that.