Change parameter order to match convention
Convetion for parameter order is `EXPECT_EQ(expected, actual)`.

Following convention can result in better error messages in the case of
failure. The error message text may list or imply which value was
expected versus actual.

Additionally, expected values tend to be more regular in length than
expressions used for actual values, so the code tends to line up nicer
in that order.