Update makefile to add library stdc++fs
This library is needed for `<filesystem>` support. It needs to be added
when compiling an executable, rather than a library. Hence this is
needed by the test project, and not the core library project.

The library needs to be included once the executable touches code that
needs `<filesystem>`, hence why ealier test project builds didn't
complain. They weren't calling parts of OP2Utility that made use of
`<filesystem>`.