Use return value struct rather than out parameter
The compiler should be able to apply Named Return Value Optimization
(NRVO) to eliminate any actual struct copy here. Internally the compiler
is free to pass a pointer into the method, much like the case of the out
parameter. This allows the out value to be constructed in place in the
stack frame of the caller.

See [copy elision](https://en.cppreference.com/w/cpp/language/copy_elision)