Add makefile commands to build Google Mock
This is only needed when Google Mock is not packaged together with
Google Test. If they are packaged together, then building Google Test
will also build Google Mock.

This update is mainly to support Ubuntu 14.04 builds, where the Apt
packages are sufficiently old that Google Test and Google Mock are not
yet bundled together.