Allow using alternate Google Mock package
This is to support using the `google-mock` Apt package on Ubuntu 14.04.
It comes with a copy of Google Test in a subfolder. If this package is
found, use it preferrentially over the separate Google Test package,
which does not include Google Mock.

Alternatively, the `libgtest-dev` Apt package can be used on newer
distributions of Ubuntu (16.04, 18.04). This package contains both
Google Test and Google Mock together, and places them in standard
locations.