Replace Tag alias with Tag class
This is a bit of an experiment to see how a simple Tag alias compares to
a full Tag class. The Tag class may be considered a bit more pure in
that it doesn't pollute the global namespace with methods that might run
on an `std::array<char, 4>` which was not intended to be a `Tag`. It
also allows for a few extra conveniences, such as copy assignment from a
string literal, and easy conversion to a `std::string`.