Access archives with full filename path
Stripping the path component here is questionable. It prevents the
archives from having subfolders contained within them. Though the
concept of a subfolder within an archive is an artificial construct. In
reality, any string is a potentially unique file identifier, with no
special interpretation for "/". Such interpretation would only be
relevant when extracting to disk. For example, an archive might contain
a filename of "C:/Some/Long/String/To/A/File.txt", where the entire
string is interpreted simply as a filename in a flat filesystem, without
any concept of drives or path components. Trying to strip a path
component would be making an assumption about the namespace within an
archive.

From another perspective, this code is after checking the disk for a
filename, where paths are actually properly relevant. Any path
manipulation should be done before raw file access code.