Add HasRootComponent and MakeAbsolute
I opted to put the relativeRootPath parameter second in case we someday
decide to default it to the current working directory. That may be a bit
difficult in the current version of C++, without using a sentinel value,
so I've left that out for now.

I've left behavior unspecified when a single root element is specified
on Windows, such as "C:", or "/".