Allow ResourceManager to handle absolute paths
This change may be a bit contentious. If ResourceManager were to be used
in a security restricting context, it may be desireable to simply
disallow absolute paths. That is however not a current design goal of
ResourceManager. As such, this seems like the proper way to handle both
absolute and relative paths.