Add DynamicMemoryWriter class
This writes to a dynamically allocated and resizable memory buffer.
After writing, the `GetReader()` method will return a `MemoryReader`
object with a view of the memory.

This class can be used to aid writing test code. Writing can be done
in-memory, and then verified with the associated `MemoryReader` object.
There is no need to write to a slow disk, nor any need to clean up
temporary files. This makes writing test code simpler and faster.