Add constexpr-if check when writing containers
If due to template parameters, the max possible container size can't
overflow the size field being written to disk, there is no needs to do a
runtime overflow check.

This is in contrast to the reverse direction of bounds checking the max
container size when reading into memory, as the container may have a
lower theoretical maximum, which might not be a `constexpr` value. See
Issue #275 for details.